Preserve public import paths
Moving the code to the `addon-test-support` tree means that all
the modules in it would have export paths starting with
`ember-cli-page-object/test-helpers` which would break a backword compatibility.
In case if we funnel all the `addon-test-support` contents to the custom
destination we would loose an ability to have valid test coverage reports.

In order to work-around it `--re-export/` directory was introduced. It's
responsible for keeping one linear re-export modules, namely `index.js`,
`extend.js` and `macros.js`. This files are funneled to the custom
`/ember-cli-page-object` destination during the build. This allows us
to leave public import paths unchanged and keep test coverage working
at the same time.